Remove coordinator (plugin) (#1404)
* - Removed coordinator

* - Removed migrator but kept receipt validation

* - Added whiteflag and extended protocol parameters to INX implementation

* - Updated go.mod

* - go mod updates

* - Running private tangle with inx-coordinator (local docker build for now)

* - Use uint32 for the whiteflag quorum api endpoint

* - Removed NewUTXOOutput and NewUTXOSpent events

- Added TreasuryMutated event

* - Implemented INX ListenToTreasuryUpdates

* - go mod tidy after rebase

* - Updated iota.go/v2 and renamed iotago.Receipt to iotago.ReceiptMilestoneOpt

* - Added verification that the last milestone ID is in the past cone when running whiteflag

- Changed whiteflag to return multiple merkle proofs
- Sending milestone ID with the milestone over INX

* - Updated inx and iota.go/v3

* - Fixed tests (integration tests still pending inx-coordinator release to be adapted)

* - Refactored IssueAndConfirmMilestoneOnTips function in the test environment into two functions, one to issue milestones and one to confirm milestones

* - Adapted white flag to not confirm the milestone message itself

- Mark all valid milestone reattachments as milestones in the metadata
- Added new whiteflag tests to handle milestone reattachments and no valid milestone past-cone

* - Fixed faucet tests checking below max depth

* - Adapted URTS test to account the milestone message itself is not referenced (i.e. ycri=ocri=milestoneIndex)

* - Updated to latest milestone payload with renamed fields and uint32 timestamp

* - Fixed tests and re-generated integration tests snapshots

* - Adapted integration-tests to use inx-coordinator

- Updated private_tangle to use the pre-built inx-coordinator docker image

* - Removed commented out code

* - Updated to latest inx-coordinator

* - Fixed integration test config params

* - Also pull indexer image

* Update core/protocfg/core.go


* Update pkg/whiteflag/white_flag.go


* - review fixes

* Update pkg/whiteflag/white_flag.go


* Update plugins/inx/server_milestones.go


* Update plugins/inx/server_milestones.go


* Update pkg/model/milestonemanager/milestone_manager.go


* - More review fixes

* - Moved whiteflag API endpoint to the core rest api

* - Updated iota.go/v3 and inx-coordinator to latest versions

* - Review fixes

* - Review fixes

+++ pkg/model/utxo/milestone_diff_test.go
@@ -17,7 +17,7 @@ import (
 
 func TestSimpleMilestoneDiffSerialization(t *testing.T) {
 	milestoneIndex := milestone.Index(255975)
-	milestoneTimestamp := rand.Uint64()
+	milestoneTimestamp := rand.Uint32()
 
 	outputID := utils.RandOutputID()
 	messageID := utils.RandMessageID()
@@ -64,7 +64,7 @@ func TestTreasuryMilestoneDiffSerialization(t *testing.T) {
 	address := utils.RandAddress(iotago.AddressEd25519)
 	amount := uint64(235234)
 	msIndex := utils.RandMilestoneIndex()
-	msTimestamp := rand.Uint64()
+	msTimestamp := rand.Uint32()
 	iotaOutput := &iotago.BasicOutput{
 		Amount: amount,
 		Conditions: iotago.UnlockConditions{
@@ -79,7 +79,7 @@ func TestTreasuryMilestoneDiffSerialization(t *testing.T) {
 	copy(transactionID[:], utils.RandBytes(iotago.TransactionIDLength))
 
 	milestoneIndex := milestone.Index(255975)
-	milestoneTimestamp := rand.Uint64()
+	milestoneTimestamp := rand.Uint32()
 
 	spent := NewSpent(output, transactionID, milestoneIndex, milestoneTimestamp)
 
@@ -138,7 +138,7 @@ func TestMilestoneDiffSerialization(t *testing.T) {
 	}
 
 	msIndex := milestone.Index(756)
-	msTimestamp := rand.Uint64()
+	msTimestamp := rand.Uint32()
 
 	spents := Spents{
 		RandUTXOSpent(outputs[3], msIndex, msTimestamp),
